Use namespaced packages instead of plain structure
This change helps to deliver the code in a way, where we can
test and develop new features without affecting customers experience.

This usage you can typically see with SDKs developed with an intention
of continuos usage, where a cost of a mistake is quite high. There are
also some language (Clojure for instance) do this by default.

The idea is to have a main package, let us say `volur.sdk` which
has mutiple versions of a client `VolurClient`. We might want to
change the implementation of the client while keeping the current one.

With this it could be easily achieved by adding a new namespaced
package `volur.sdk.v1alpha2` or `volur.sdk.v2` in a case of a breaking
change. The main package `volur.sdk` will still point to a present
implementation (ideally stable one).